% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_not_reassign.R
\name{.delete}
\alias{.delete}
\title{Delete a cached object}
\usage{
.delete(obj)
}
\arguments{
\item{obj}{object}
}
\value{
NULL. Deletes \code{obj}, if it exists, and removes saved file, if it
  exists.
}
\description{
Delete a cached object
}
