% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_not_reassign.R
\name{.rename}
\alias{.rename}
\title{Rename a cached object}
\usage{
.rename(obj, new_name)
}
\arguments{
\item{obj}{object}

\item{new_name}{character}
}
\value{
Boolean.
}
\description{
Rename a cached object
}
