% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_not_reassign.R
\name{\%<~\%}
\alias{\%<~\%}
\title{Cached assignment}
\usage{
lhs \%<~\% rhs
}
\arguments{
\item{lhs, rhs}{left and right hand side of an assignment}
}
\value{
NULL. Assigns to \code{lhs} and possibly saves an RDS to
  \code{save_folder}.
}
\description{
Assignment that (i) doesn't reassign if the object already exists in the
current environment, (ii) saves \code{lhs} to the folder set by
\code{\link{set_save_folder}}, and (iii) loads from file when \code{lhs.RDS}
exists in the save file. Intended only to be used in the global environment.
}
\examples{
 \dontrun{
   folder <- "."
   set_save_folder(folder)
   a \%<~\% 1
   a
   rm(a)
   a \%<~\% 2
   a
   .delete(a)
   a \%<~\% 2
   a
   .rename(a, "b")
   a \%<~\% 1
   a
   b
   list.files(folder)
 }
}
