% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_not_reassign.R
\name{set_save_folder}
\alias{set_save_folder}
\title{Set save folder for caching}
\usage{
set_save_folder(folder = getwd())
}
\arguments{
\item{folder}{string}
}
\value{
NULL, sets global variable \code{save_folder}.
}
\description{
Set save folder for caching
}
